fix(python3): add `pyproject.toml` to include NumPy during installation
Because `pip install` will run in an isolated environment by default, we
(sometimes) need to explicitly declare which modules are to be included,
and because NumPy exposes its own C API, we need to be able to query the
module at installation time. We can tell Pip to include it by adding
`numpy` as a dependency of the build system.