fix(iDISF_py3.c): convert and cast arrays to `PyArrayObject`
Some functions in newer versions of the NumPy C API can no longer accept
`PyObject *` parameters, only `PyArrayObject *`. Because the existing
parameters are already guaranteed to be NumPy arrays due to existing
validations and conversions, we can simply redeclare them as
`PyArrayObject *` when possible, and statically cast them when not.